+++ CHANGELOG.md
@@ -38,6 +38,7 @@
   - [Spread operator](https://caniuse.com/?search=spread%20operator)
   - [destructuring](https://caniuse.com/?search=destructuring)
   - [default parameters](https://caniuse.com/?search=default%20parameters)
+  - [object spread](https://caniuse.com/?search=Object%20spread)
 
 
 <!-- /CHANGELOG NEWEST -->